requirements.txt
# ByteMe Enterprise Assistant - Requirements
# Core Framework
streamlit>=1.28.0

# LangChain & LangGraph
langchain>=0.1.0
langchain-community>=0.0.10
langchain-huggingface>=0.0.1
langchain-groq>=0.0.1
langchain-experimental>=0.0.40
langgraph>=0.0.20

# Vector Database
chromadb>=0.4.0

# PostgreSQL (Neon DB) - User Auth & Long-term Memory
psycopg2-binary>=2.9.0

# Embeddings & ML
sentence-transformers>=2.2.0
torch>=2.0.0
transformers>=4.30.0

# PDF Processing
pymupdf>=1.23.0

# Redis (Short-term memory)
redis>=5.0.0

# Optional: Vision Model (requires GPU)
# colpali-engine>=0.1.0
# bitsandbytes>=0.41.0
# peft>=0.6.0

# Utilities
numpy>=1.24.0
pandas>=2.0.0
pyzbar>=0.1.9
Pillow>=10.0.0
tqdm>=4.65.0

# Environment
python-dotenv>=1.0.0
